local-llm/requirements.txt
httpx==0.27.2
python-socketio==5.11.4
asyncio==3.4.3
pydantic==2.9.2
tiktoken==0.7.0
aiofiles==24.1.0
openai-whisper
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
transformers==4.44.2
accelerate==0.34.2
diffusers==0.30.3
